+++ lib/selenium_fury/page_object/element_types/drop_down_element.rb
@@ -13,12 +13,14 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
-
-require 'selenium_fury'
-
-require 'test_page/test_page_custom_generator_configuration'
-require 'test_page/test_page'
-
-TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"
-
-include SeleniumFury::Utilities
+module SeleniumFury
+  module PageObject
+    module ElementTypes
+      # This class represents a drop down element on a web page.
+      class DropDownElement < GenericElement
+        include ElementHelpers::DropDownElementHelper
+        include ElementHelpers::SelectableElementHelper
+      end # DropDownElement
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
