+++ lib/selenium_fury/page_object/element_types/element_helpers/checkbox_element_helper.rb
@@ -13,12 +13,18 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
+module SeleniumFury
+  module PageObject
+    module ElementTypes
+      module ElementHelpers
+        # Defines what we can do to a checkbox element
+        module CheckboxElementHelper
+          def checked(be_selected=true)
+            select unless be_selected == selected?
+          end
+        end #CheckboxElementHelper
+      end # ElementHelpers
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
 
-require 'selenium_fury'
-
-require 'test_page/test_page_custom_generator_configuration'
-require 'test_page/test_page'
-
-TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"
-
-include SeleniumFury::Utilities
