+++ lib/selenium_fury/page_object/element_types/element_helpers/image_element_helper.rb
@@ -13,12 +13,22 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
+module SeleniumFury
+  module PageObject
+    module ElementTypes
+      module ElementHelpers
+        # Defines what we can do to an image element
+        module ImageElementHelper
+          def text
+            el.attribute('alt')
+          end
 
-require 'selenium_fury'
+          def source
+            el.attribute('src')
+          end
+        end # ImageElementHelper
+      end # ElementHelpers
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
 
-require 'test_page/test_page_custom_generator_configuration'
-require 'test_page/test_page'
-
-TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"
-
-include SeleniumFury::Utilities
