+++ lib/selenium_fury/page_object/element_types/element_helpers/text_input_element_helper.rb
@@ -13,12 +13,25 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
+module SeleniumFury
+  module PageObject
+    module ElementTypes
+      module ElementHelpers
+        # Defines what we can do to a text input element
+        module TextInputElementHelper
+          def send_keys(text)
+            raise "Locator at #{location} can not be interacted with" unless visible?
+            el.clear
+            el.send_keys(text)
+          end
 
-require 'selenium_fury'
+          def send_keys!(text)
+            raise "Locator at #{location} can not be interacted with" unless visible?
+            el.send_keys(text)
+          end
+        end # TextInputElementHelper
+      end # ElementHelpers
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
 
-require 'test_page/test_page_custom_generator_configuration'
-require 'test_page/test_page'
-
-TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"
-
-include SeleniumFury::Utilities
