+++ lib/selenium_fury/page_object/element_types/selectable_element.rb
@@ -13,22 +13,13 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
-
 module SeleniumFury
-  module SeleniumWebDriver
-    module CreateSeleniumWebDriver
-
-      #@return [Selenium::WebDriver::Driver]
-      def driver
-        return @driver
-      end
-
-      # @param url [string]
-      # @return [Selenium::WebDriver::Driver]
-      def launch_web_driver url=nil
-        @driver = Selenium::WebDriver.for :chrome
-        @driver.navigate.to url unless url.nil?
-      end
-    end
-  end
-end
+  module PageObject
+    module ElementTypes
+      # This class represents a selectable element on a web page.
+      class SelectableElement < GenericElement
+        include ElementHelpers::SelectableElementHelper
+      end # SelectableElement
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
