+++ lib/selenium_fury/page_object/element_types/submit_element.rb
@@ -13,12 +13,13 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
-
-require 'selenium_fury'
-
-require 'test_page/test_page_custom_generator_configuration'
-require 'test_page/test_page'
-
-TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"
-
-include SeleniumFury::Utilities
+module SeleniumFury
+  module PageObject
+    module ElementTypes
+      # This class represents a submit element on a web page.
+      class SubmitElement < GenericElement
+        include ElementHelpers::SelectableElementHelper
+      end # SubmitElement
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
