+++ lib/selenium_fury/page_object/element_types/text_element.rb
@@ -13,12 +13,12 @@
 # * See the License for the specific language governing permissions and
 # * limitations under the License.
 # */
-
-require 'selenium_fury'
-
-require 'test_page/test_page_custom_generator_configuration'
-require 'test_page/test_page'
-
-TEST_PAGE_URL="file://#{File.dirname(__FILE__)}/test_page/test_page.html"
-
-include SeleniumFury::Utilities
+module SeleniumFury
+  module PageObject
+    module ElementTypes
+      # This class represents a text element on a web page.
+      class TextElement < GenericElement
+      end # TextElement
+    end # ElementTypes
+  end # PageObject
+end # SeleniumFury
