+++ spec/selenium_fury/page_object/element_types/element_helpers/text_input_element_helper_spec.rb
@@ -1,0 +1,27 @@
+require 'spec_helper'
+
+describe ElementTypes::ElementHelpers::TextInputElementHelper do
+
+  let(:test_page) { TestPage.new(driver) }
+
+  before(:all) do
+    launch_web_driver TEST_PAGE_URL
+  end
+
+  after(:all) do
+    stop_web_driver
+  end
+
+  it 'should clear and write text' do
+    text = 'Hey buddy'
+    test_page.textarea_element.send_keys(text)
+    test_page.textarea_element.value.should == text
+  end
+
+  it 'should write text without clearing' do
+    existing_text = test_page.textarea_element.value
+    new_text = 'Hey buddy'
+    test_page.textarea_element.send_keys!(new_text)
+    test_page.textarea_element.value.should == existing_text+new_text
+  end
+end
